+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.35.0 - (March 25, 2022)
+
 * Added
   * Added new prop `columnHighlightId` to highlight a column with style intended for use in timelines